The explanation for a particular answer can be very long. We need a kind of "toggle" box for that in which we can elaborate on certain aspects.
Since this was not a feature, it was added to the Token Class as a %TB_B% (toggle block begin) and %TB_E% (toggle block end) token (lib/Wpsqt/Tokens.php).
A javascript was added to actualy toggle the div's generated by set tokens (js/site.js)
The doReplacement method was used in "pages/site/quiz/section.php" to actualy parse the tokens when used.
As a bonus all other tokens can now be used here.