Added a check of AutoMarkWhenFreetext option.
When not 'no' (values 'yes - ..') it will disable the check whether or not to automark and show a result on the "quiz review" page.

Values: 'no' and 'yes - include freetext' will mark freetext questions as 'incorrect',
Value: 'yes - exclude freetext' will ignore the freetext questions' and not add them to the $totalPoints (which results in a higher score)

examples:
2 freetext, 1 multiple, 3 single
all muliple/single questions answered correctly.
Result for setting the option AutoMarkWhenFreetext to:
 - no				"Total Points quiz can't be auto marked"
 - yes include freetext		"Total Points 4 correct out of 6" (66%)
 - yes exclude freetext		"Total Points 4 correct out of 4" (100%)

which can make a difference whether the quiz result was marked as a fail.