Only allow one quiz to be running at a time
Previously when quizzes were inserted into posts and multiple posts were
displayed at once, progressing in one quiz would also cause the other
quizzes to progress.  This would cause many errors.

The quiz name has been added to the form and this is checked against the
current quiz to ensure that other quizzes are not progressed as well.